Fix building the dump lookahead data feature

+++ src/api/mod.rs
@@ -1831,8 +1831,8 @@ impl<T: Pixel> ContextInner<T> {
         let data = &fi.block_importances;
         use byteorder::{NativeEndian, WriteBytesExt};
         let mut buf = vec![];
-        buf.write_u64::<NativeEndian>(fi.h_in_imp_b).unwrap();
-        buf.write_u64::<NativeEndian>(fi.w_in_imp_b).unwrap();
+        buf.write_u64::<NativeEndian>(fi.h_in_imp_b as u64).unwrap();
+        buf.write_u64::<NativeEndian>(fi.w_in_imp_b as u64).unwrap();
         for y in 0..fi.h_in_imp_b {
           for x in 0..fi.w_in_imp_b {
             let importance = data[y * fi.w_in_imp_b + x];
